Update pgeocode.py
Simple modification at line 254 that should solve issue #32 and issue #55. latitude was misspelled as  "lattitude". With this misspelling the field latitude *does* figure in the valid_keys list and thus the average value of the latitude will be overwritten in line 260 with the first value seen in the source database and thus the coordinates are slightly shifted.